renaming to phase out old commandHandlers architecture
